graph/requirements.txt
requests
google-cloud-storage
matplotlib
flask